+++ src/main/java/com/wztwzt/ae2_qof/hatch/adaptive/AdaptiveNetLaserHatch.java
@@ -28,6 +28,7 @@ import gregtech.api.interfaces.tileentity.IGregTechTileEntity;
 import gregtech.api.metatileentity.MetaTileEntity;
 import gregtech.api.metatileentity.implementations.MTEHatchEnergy;
 import gregtech.api.render.TextureFactory;
+import gregtech.api.util.GTUtility;
 
 import gregtech.common.misc.WirelessNetworkManager;
 
@@ -130,7 +131,12 @@ public class AdaptiveNetLaserHatch extends MTEHatchEnergy {
             gregtech.api.metatileentity.MetaTileEntity mte = (gregtech.api.metatileentity.MetaTileEntity) aBase.getMetaTileEntity();
             if (mte != null) {
                 net.minecraft.item.ItemStack stack = mte.getStackForm(1L);
-                helper.setCachedInfo((short) stack.getItemDamage(), stack.getDisplayName());
+                if (stack != null) {
+                    helper.setCachedInfo((short) stack.getItemDamage(), stack.getDisplayName());
+                } else {
+                    helper.setCachedInfo((short) -1,
+                        net.minecraft.util.StatCollector.translateToLocal(helper.getHatchType().getTranslationKey()));
+                }
             }
             lastStoredEU = aBase.getStoredEU();
             if (helper.isBound()) {
@@ -181,10 +187,10 @@ public class AdaptiveNetLaserHatch extends MTEHatchEnergy {
 
         column.child(new TextWidget<>(IKey.dynamic(() -> {
             int v = helper.getCurrentVoltageTier();
-            String tierName = HatchType.getTierName(v);
+            String tierName = GTUtility.getColoredTierNameFromTier((byte) v);
             return EnumChatFormatting.AQUA
                 + StatCollector.translateToLocal("ae2_qof.gui.adaptive_hatch.voltage")
-                + " " + EnumChatFormatting.WHITE + tierName + " (" + V[v] + " EU/t)";
+                + " " + tierName + " (" + V[v] + " EU/t)";
         })).size(260, 14));
 
         column.child(new TextWidget<>(IKey.dynamic(() -> {
@@ -253,7 +259,7 @@ public class AdaptiveNetLaserHatch extends MTEHatchEnergy {
         super.getWailaBody(itemStack, currenttip, accessor, config);
         try {
             int v = helper.getCurrentVoltageTier();
-            String tierName = HatchType.getTierName(v);
+            String tierName = GTUtility.getColoredTierNameFromTier((byte) v);
             currenttip.add(EnumChatFormatting.AQUA + "V:" + tierName + " (" + V[v] + ") | A:" + helper.getCurrentAmps());
             java.math.BigInteger gridEU = helper.isBound() && helper.getNetworkOwner() != null
                 ? WirelessNetworkManager.getUserEU(helper.getNetworkOwner()) : java.math.BigInteger.ZERO;